+++ src/main/kotlin/no/nav/melosys/skjema/integrasjon/arbeidsgiver/ArbeidsgiverNotifikasjonConsumer.kt
@@ -1,24 +1,18 @@
 package no.nav.melosys.skjema.integrasjon.arbeidsgiver
 
 import io.github.oshai.kotlinlogging.KotlinLogging
+import no.nav.melosys.skjema.integrasjon.arbeidsgiver.dto.BeskjedRequest
 import no.nav.melosys.skjema.integrasjon.arbeidsgiver.dto.GraphQLRequest
 import no.nav.melosys.skjema.integrasjon.arbeidsgiver.dto.GraphQLResponse
 import no.nav.melosys.skjema.integrasjon.arbeidsgiver.dto.NyBeskjedResponse
+import no.nav.melosys.skjema.integrasjon.arbeidsgiver.dto.NyBeskjedVariables
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.core.io.ClassPathResource
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.bodyToMono
-import java.util.UUID
 
 private val log = KotlinLogging.logger { }
-
-data class BeskjedRequest(
-    val virksomhetsnummer: String,
-    val tekst: String,
-    val lenke: String,
-    val eksternId: String = UUID.randomUUID().toString(),
-)
 
 @Component
 class ArbeidsgiverNotifikasjonConsumer(
@@ -34,16 +28,18 @@ class ArbeidsgiverNotifikasjonConsumer(
     fun opprettBeskjed(request: BeskjedRequest): String {
         log.info { "Oppretter ny beskjed for virksomhet ${request.virksomhetsnummer} med eksternId ${request.eksternId}" }
 
+        val variables = NyBeskjedVariables(
+            request.eksternId,
+            request.virksomhetsnummer,
+            request.lenke,
+            request.tekst,
+            merkelapp,
+            ressursId,
+        )
+
         val graphQLRequest = GraphQLRequest(
             query = nyBeskjedMutation,
-            variables = mapOf(
-                "eksternId" to request.eksternId,
-                "virksomhetsnummer" to request.virksomhetsnummer,
-                "lenke" to request.lenke,
-                "tekst" to request.tekst,
-                "merkelapp" to merkelapp,
-                "ressursId" to ressursId,
-            )
+            variables = variables
         )
 
         val response = arbeidsgiverNotifikasjonClient.post()
